// library/jsx-server/jsx-runtime.backup.d.ts
declare namespace JSX {
  // Our JSX element cannot be simple string otherwise we can't escape properly HTML.
  export type Child = string | number | boolean | null | undefined | JSXElement
  export type JSXElement =
    | { type: "rawHtml"; children: string }
    | { type: "fragment"; children: Child[] }
    | {
        type: "element"
        tag: keyof JSX.IntrinsicElements
        attributes: Record<string, any>
        children: Child[]
      }
// ...
  interface HTMLAttributes extends AriaAttributes, DOMAttributes {
    // Standard HTML Attributes
    accesskey?: string | undefined | null
    autocapitalize?: string | undefined | null
    autofocus?: boolean | string | undefined | null
    class?: string | undefined | null
    contenteditable?:
      | "true"
      | "false"
      | boolean
      | "inherit"
      | string
      | undefined
      | null
    dir?: string | undefined | null
    draggable?: "true" | "false" | boolean | undefined | null
    enterkeyhint?:
      | "enter"
      | "done"
      | "go"
      | "next"
      | "previous"
      | "search"
      | "send"
      | undefined
      | null
    exportparts?: string | undefined | null
    hidden?: boolean | string | undefined | null
    id?: string | undefined | null
    inert?: boolean | string | undefined | null
    inputmode?:
      | "none"
      | "text"
      | "tel"
      | "url"
      | "email"
      | "numeric"
      | "decimal"
      | "search"
      | undefined
      | null
    is?: string | undefined | null

    // Microdata API
    itemid?: string | undefined | null
    itemprop?: string | undefined | null
    itemref?: string | undefined | null
    itemscope?: boolean | string | undefined | null
    itemtype?: string | undefined | null

    lang?: string | undefined | null
    part?: string | undefined | null
    popover?: boolean | string | undefined | null
    slot?: string | undefined | null
    spellcheck?: "true" | "false" | boolean | undefined | null
    style?: string | undefined | null
    tabindex?: number | string | undefined | null
    title?: string | undefined | null
    translate?: "yes" | "no" | "" | undefined | null

    // <command>, <menuitem>
    radiogroup?: string | undefined | null

    // WAI-ARIA
    role?: AriaRole | undefined | null

    // RDFa Attributes
    about?: string | undefined | null
    datatype?: string | undefined | null
    inlist?: any
    prefix?: string | undefined | null
    property?: string | undefined | null
    resource?: string | undefined | null
    typeof?: string | undefined | null
    vocab?: string | undefined | null

    // Non-standard Attributes
    contextmenu?: string | undefined | null // Obsolete
    autosave?: string | undefined | null // Apple exclusive
    color?: string | undefined | null
    results?: number | string | undefined | null
    security?: string | undefined | null
    unselectable?: "on" | "off" | undefined | null // Internet Explorer

    // Allow data- attribute
    [key: `data-${string}`]: any
  }

  type HTMLAttributeReferrerPolicy =
    | ""
    | "no-referrer"
    | "no-referrer-when-downgrade"
    | "origin"
    | "origin-when-cross-origin"
    | "same-origin"
    | "strict-origin"
    | "strict-origin-when-cross-origin"
    | "unsafe-url"

  type HTMLAttributeAnchorTarget =
    | "_self"
    | "_blank"
    | "_parent"
    | "_top"
    | (string & {})

  interface AnchorHTMLAttributes extends HTMLAttributes {
    download?: string | boolean | undefined | null
    href?: string | URL | undefined | null
    hreflang?: string | undefined | null
    media?: string | undefined | null
    name?: string | undefined | null
    ping?: string | undefined | null
    rel?: string | undefined | null
    target?: HTMLAttributeAnchorTarget | undefined | null
    type?: string | undefined | null
    referrerpolicy?: HTMLAttributeReferrerPolicy | undefined | null
  }

  interface AudioHTMLAttributes extends MediaHTMLAttributes {}

  interface AreaHTMLAttributes extends HTMLAttributes {
    alt?: string | undefined | null
    coords?: string | undefined | null
    download?: any
    href?: string | undefined | null
    hreflang?: string | undefined | null
    media?: string | undefined | null
    referrerpolicy?: HTMLAttributeReferrerPolicy | undefined | null
    rel?: string | undefined | null
    shape?: string | undefined | null
    target?: string | undefined | null
  }

  interface BaseHTMLAttributes extends HTMLAttributes {
    href?: string | undefined | null
    target?: string | undefined | null
  }

  interface BlockquoteHTMLAttributes extends HTMLAttributes {
    cite?: string | undefined | null
  }

  interface ButtonHTMLAttributes extends HTMLAttributes {
    autocomplete?: string | undefined | null
    disabled?: boolean | string | undefined | null
    form?: string | undefined | null
    formaction?: string | undefined | null
    formenctype?: string | undefined | null
    formmethod?: string | undefined | null
    formnovalidate?: boolean | string | undefined | null
    formtarget?: string | undefined | null
    name?: string | undefined | null
    type?: "submit" | "reset" | "button" | undefined | null
    value?: string | string[] | number | undefined | null
    popovertarget?: string | undefined | null
    popovertargetaction?: "hide" | "show" | "toggle" | undefined | null
  }

  interface CanvasHTMLAttributes extends HTMLAttributes {
    height?: number | string | undefined | null
    width?: number | string | undefined | null
  }

  interface ColHTMLAttributes extends HTMLAttributes {
    span?: number | string | undefined | null
    width?: number | string | undefined | null
  }

  interface ColgroupHTMLAttributes extends HTMLAttributes {
    span?: number | string | undefined | null
  }

  interface DataHTMLAttributes extends HTMLAttributes {
    value?: string | string[] | number | undefined | null
  }

  interface DetailsHTMLAttributes extends HTMLAttributes {
    open?: boolean | string | undefined | null
    name?: string | undefined | null
  }

  interface DelHTMLAttributes extends HTMLAttributes {
    cite?: string | undefined | null
    datetime?: string | undefined | null
  }

  interface DialogHTMLAttributes extends HTMLAttributes {
    open?: boolean | string | undefined | null
  }

  interface EmbedHTMLAttributes extends HTMLAttributes {
    height?: number | string | undefined | null
    src?: string | undefined | null
    type?: string | undefined | null
    width?: number | string | undefined | null
  }

  interface FieldsetHTMLAttributes extends HTMLAttributes {
    disabled?: boolean | string | undefined | null
    form?: string | undefined | null
    name?: string | undefined | null
  }

  interface FormHTMLAttributes extends HTMLAttributes {
    "accept-charset"?: string | undefined | null
    action?: string | undefined | null
    autocomplete?: string | undefined | null
    autocorrect?: string | undefined | null
    enctype?: string | undefined | null
    method?: string | undefined | null
    name?: string | undefined | null
    novalidate?: boolean | string | undefined | null
    target?: string | undefined | null
  }

  interface HtmlHTMLAttributes extends HTMLAttributes {
    manifest?: string | undefined | null
  }
// ...
  interface InsHTMLAttributes extends HTMLAttributes {
    cite?: string | undefined | null
    datetime?: string | undefined | null
  }
// ...
  interface LabelHTMLAttributes extends HTMLAttributes {
    form?: string | undefined | null
    for?: string | undefined | null
  }

  interface LiHTMLAttributes extends HTMLAttributes {
    value?: string | number | undefined | null
  }
// ...
  interface IntrinsicElements {
    // HTML
    a: AnchorHTMLAttributes
    abbr: HTMLAttributes
    address: HTMLAttributes
    area: AreaHTMLAttributes
    article: HTMLAttributes
    aside: HTMLAttributes
    audio: AudioHTMLAttributes
    b: HTMLAttributes
    base: BaseHTMLAttributes
    bdi: HTMLAttributes
    bdo: HTMLAttributes
    big: HTMLAttributes
    blockquote: BlockquoteHTMLAttributes
    body: HTMLAttributes
    br: HTMLAttributes
    button: ButtonHTMLAttributes
    canvas: CanvasHTMLAttributes
    caption: HTMLAttributes
    cite: HTMLAttributes
    code: HTMLAttributes
    col: ColHTMLAttributes
    colgroup: ColgroupHTMLAttributes
    data: DataHTMLAttributes
    datalist: HTMLAttributes
    dd: HTMLAttributes
    del: DelHTMLAttributes
    details: DetailsHTMLAttributes
    dfn: HTMLAttributes
    dialog: DialogHTMLAttributes
    div: HTMLAttributes
    dl: HTMLAttributes
    dt: HTMLAttributes
    em: HTMLAttributes
    embed: EmbedHTMLAttributes
    fieldset: FieldsetHTMLAttributes
    figcaption: HTMLAttributes
    figure: HTMLAttributes
    footer: HTMLAttributes
    form: FormHTMLAttributes
    h1: HTMLAttributes
    h2: HTMLAttributes
    h3: HTMLAttributes
    h4: HTMLAttributes
    h5: HTMLAttributes
    h6: HTMLAttributes
    head: HTMLAttributes
    header: HTMLAttributes
    hgroup: HTMLAttributes
    hr: HTMLAttributes
    html: HtmlHTMLAttributes
    i: HTMLAttributes
    iframe: IframeHTMLAttributes
    img: ImgHTMLAttributes
    input: InputHTMLAttributes
    ins: InsHTMLAttributes
    kbd: HTMLAttributes
    keygen: KeygenHTMLAttributes
    label: LabelHTMLAttributes
    legend: HTMLAttributes
    li: LiHTMLAttributes
    link: LinkHTMLAttributes
    main: HTMLAttributes
    map: MapHTMLAttributes
    mark: HTMLAttributes
    menu: MenuHTMLAttributes
    menuitem: HTMLAttributes
    meta: MetaHTMLAttributes
    meter: MeterHTMLAttributes
    nav: HTMLAttributes
    noindex: HTMLAttributes // https://en.wikipedia.org/wiki/Noindex#%3Cnoindex%3E_tag
    noscript: HTMLAttributes
    object: ObjectHTMLAttributes
    ol: OlHTMLAttributes
    optgroup: OptgroupHTMLAttributes
    option: OptionHTMLAttributes
    output: OutputHTMLAttributes
    p: HTMLAttributes
    param: ParamHTMLAttributes
    picture: HTMLAttributes
    pre: HTMLAttributes
    progress: ProgressHTMLAttributes
    q: QuoteHTMLAttributes
    rp: HTMLAttributes
    rt: HTMLAttributes
    ruby: HTMLAttributes
    s: HTMLAttributes
    samp: HTMLAttributes
    slot: SlotHTMLAttributes
    script: ScriptHTMLAttributes
    section: HTMLAttributes
    select: SelectHTMLAttributes
    small: HTMLAttributes
    source: SourceHTMLAttributes
    span: HTMLAttributes
    strong: HTMLAttributes
    style: StyleHTMLAttributes
    sub: HTMLAttributes
    summary: HTMLAttributes
    sup: HTMLAttributes
    table: TableHTMLAttributes
    tbody: HTMLAttributes
    td: TdHTMLAttributes
    template: HTMLAttributes
    textarea: TextareaHTMLAttributes
    tfoot: HTMLAttributes
    th: ThHTMLAttributes
    thead: HTMLAttributes
    time: TimeHTMLAttributes
    title: HTMLAttributes
    tr: HTMLAttributes
    track: TrackHTMLAttributes
    u: HTMLAttributes
    ul: HTMLAttributes
    var: HTMLAttributes
    video: VideoHTMLAttributes
    wbr: HTMLAttributes

    // SVG
    svg: SVGAttributes
    animate: SVGAttributes
    circle: SVGAttributes
    clipPath: SVGAttributes
    defs: SVGAttributes
    desc: SVGAttributes
    ellipse: SVGAttributes
    feBlend: SVGAttributes
    feColorMatrix: SVGAttributes
    feComponentTransfer: SVGAttributes
    feComposite: SVGAttributes
    feConvolveMatrix: SVGAttributes
    feDiffuseLighting: SVGAttributes
    feDisplacementMap: SVGAttributes
    feDistantLight: SVGAttributes
    feFlood: SVGAttributes
    feFuncA: SVGAttributes
    feFuncB: SVGAttributes
    feFuncG: SVGAttributes
    feFuncR: SVGAttributes
    feGaussianBlur: SVGAttributes
    feImage: SVGAttributes
    feMerge: SVGAttributes
    feMergeNode: SVGAttributes
    feMorphology: SVGAttributes
    feOffset: SVGAttributes
    fePointLight: SVGAttributes
    feSpecularLighting: SVGAttributes
    feSpotLight: SVGAttributes
    feTile: SVGAttributes
    feTurbulence: SVGAttributes
    filter: SVGAttributes
    foreignObject: SVGAttributes
    g: SVGAttributes
    image: SVGAttributes
    line: SVGAttributes
    linearGradient: SVGAttributes
    marker: SVGAttributes
    mask: SVGAttributes
    metadata: SVGAttributes
    path: SVGAttributes
    pattern: SVGAttributes
    polygon: SVGAttributes
    polyline: SVGAttributes
    radialGradient: SVGAttributes
    rect: SVGAttributes
    stop: SVGAttributes
    switch: SVGAttributes
    symbol: SVGAttributes
    text: SVGAttributes
    textPath: SVGAttributes
    tspan: SVGAttributes
    use: SVGAttributes
    view: SVGAttributes
  }
}